fix: 修复签名验证和数据库 schema 不匹配
- 修复 auth.rs: 正确收集所有查询参数用于签名验证
  (req.queries() 返回 HashMap<String, Vec<String>>)
- 修复 resource.rs: commission_rate 类型 i64 → Decimal (匹配 NUMERIC(5,4))
- 修复 migration: 更新 resources 表定义以匹配实际 schema
  - 添加 thumbnail_url, usage_guide, precautions 列
  - 修复 category_id INTEGER → BIGINT
  - 修复 download_count INTEGER → BIGINT
- 移除 auth.rs 和 signature.rs 中的调试 eprintln! 代码

+++ rsws_model/src/resource.rs
@@ -29,7 +29,7 @@ pub struct Resource {
     pub owner_type: String,
     pub provider_id: Option<i64>,
     pub supported_os: Option<serde_json::Value>,
-    pub commission_rate: i64,
+    pub commission_rate: Decimal,
     pub download_count: i64,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
